+++ app/api/events/get-information/route.ts
@@ -2,7 +2,7 @@ import { fetchEventById } from "@/app/lib/data";
 import { NextResponse } from "next/server";
 
 export async function POST(req: Request) {
-	const id = await req.json();
-	const event = await fetchEventById(id);
+	const { id }: { id: string } = await req.json();
+	const event = await fetchEventById(id);	
 	return NextResponse.json(event)
 }