+++ app/events/[id]/page.tsx
@@ -3,7 +3,7 @@ import EventInfo from "./event-info"
 
 export default function Page() {
 	return (
-		<main className='relative flex flex-col min-h-screen mx-auto p-8 pt-16 bg-gradient-to-b from-[#ffffff] via-[#d1d2d2] to-[#76d7cc] '>
+		<main className='relative flex flex-col min-h-screen mx-auto p-8 pt-16 bg-gradient-to-b from-[#f8f9fa] via-[#e9ecef] to-[#feffff]'>
 			<EventInfo />
 		</main>
 	)